+++ src/Desease/Desease.h
@@ -5,7 +5,7 @@ namespace DeseaseSpreadSimulation
 	class Desease
 	{
 	public:
-		Desease(std::string name, const int incubationPeriod, const int daysInfectious, std::pair<int, int> deseaseDurationRange, std::vector<float> mortalityByAge, std::pair<int, int> daysTillDeathRange, float spreadFactor = 1.0f);
+		Desease(std::string name, const uint16_t incubationPeriod, const uint16_t daysInfectious, std::pair<uint16_t, uint16_t> deseaseDurationRange, std::vector<float> mortalityByAge, std::pair<uint16_t, uint16_t> daysTillDeathRange, float spreadFactor = 1.0f);
 
 		// Check with ID
 		inline bool operator==(const Desease& rhs) const
@@ -14,12 +14,12 @@ namespace DeseaseSpreadSimulation
 		};
 
 		const std::string& GetDeseaseName() const;
-		int IncubationPeriod() const;
-		int DaysInfectious() const;
-		float GetMortalityByAge(int age) const;
+		uint16_t IncubationPeriod() const;
+		uint16_t DaysInfectious() const;
+		float GetMortalityByAge(uint16_t age) const;
 		float GetMortalityByAgeGroup(Age_Group age) const;
-		int GetDeseaseDuration() const;
-		int DaysTillDeath() const;
+		uint16_t GetDeseaseDuration() const;
+		uint16_t DaysTillDeath() const;
 		float GetSpreadFactor() const;
 
 		uint32_t GetID() const;
@@ -32,14 +32,14 @@ namespace DeseaseSpreadSimulation
 		const uint32_t id = 0;
 		const std::string name{};
 		// incubation period in days
-		const int incubationPeriod = 0;
+		const uint16_t incubationPeriod = 0;
 		// patient is contagious for x days after start of symptoms
-		const int daysInfectious = 0;
+		const uint16_t daysInfectious = 0;
 		// duration min, max in days
-		const std::pair<int, int> durationRange{};
+		const std::pair<uint16_t, uint16_t> durationRange{};
 		// mortality in percent from 0-1, age from 0-9, 10-19, 20-29,...., >80 years
 		const std::vector<float> mortalityByAge;
-		const std::pair<int, int> daysTillDeathRange{};
+		const std::pair<uint16_t, uint16_t> daysTillDeathRange{};
 		// In percent from 0-1 how likely it is to get infected at exposure
 		float spreadFactor = 1.0f;
 	};
